+++ coderadar-rest/src/test/java/io/reflectoring/coderadar/rest/unit/user/RegisterUserControllerTest.java
@@ -27,6 +27,6 @@ class RegisterUserControllerTest {
     ResponseEntity<IdResponse> responseEntity = testSubject.register(command);
 
     Assertions.assertEquals(HttpStatus.CREATED, responseEntity.getStatusCode());
-    Assertions.assertEquals(1L, responseEntity.getBody().getId().longValue());
+    Assertions.assertEquals(1L, responseEntity.getBody().getId());
   }
 }
